+++ components/aos_apps/aos_apps.c
@@ -14,6 +14,7 @@ void aos_apps_register_builtin(void)
         aos_app_calendar_get,
         aos_app_music_get,
         aos_app_remote_get,
+        aos_app_pcremote_get,
         aos_app_photos_get,
         aos_app_flashlight_get,
         aos_app_level_get,